:recycle: Refactor helper as common for DF and Vec
