+++ lib/red_amber/data_frame_selectable.rb
@@ -157,35 +157,6 @@ module RedAmber
 
     private
 
-    def parse_to_vector(args)
-      a = args.reduce([]) do |accum, elem|
-        accum.concat(normalize_element(elem))
-      end
-      Vector.new(a)
-    end
-
-    def normalize_element(elem)
-      case elem
-      when Numeric, String, Symbol, TrueClass, FalseClass, NilClass
-        [elem]
-      when Range
-        both_end = [elem.begin, elem.end]
-        both_end[1] -= 1 if elem.exclude_end? && elem.end.is_a?(Integer)
-
-        if both_end.any?(Integer) || both_end.all?(&:nil?)
-          if both_end.any? { |e| e&.>=(size) || e&.<(-size) }
-            raise DataFrameArgumentError, "Index out of range: #{elem} for 0..#{size - 1}"
-          end
-
-          (0...size).to_a[elem]
-        else
-          elem.to_a
-        end
-      else
-        Array(elem)
-      end
-    end
-
     def select_vars_by_keys(keys)
       if keys.one?
         key = keys[0].to_sym
